Extract command for secrets
