+++ cmd/docker-mcp/commands/root.go
@@ -14,7 +14,6 @@ import (
 	"github.com/docker/mcp-gateway/cmd/docker-mcp/internal/docker"
 	"github.com/docker/mcp-gateway/cmd/docker-mcp/oauth"
 	"github.com/docker/mcp-gateway/cmd/docker-mcp/secret-management/policy"
-	"github.com/docker/mcp-gateway/cmd/docker-mcp/secret-management/secret"
 	"github.com/docker/mcp-gateway/cmd/docker-mcp/version"
 )
 
@@ -74,7 +73,7 @@ func Root(ctx context.Context, cwd string, dockerCli command.Cli) *cobra.Command
 
 	dockerClient := docker.NewClient(dockerCli)
 
-	cmd.AddCommand(secret.NewSecretsCmd(dockerClient))
+	cmd.AddCommand(secretCommand(dockerClient))
 	cmd.AddCommand(policy.NewPolicyCmd())
 	cmd.AddCommand(oauth.NewOAuthCmd())
 	cmd.AddCommand(client.NewClientCmd(cwd))
